requirements.txt
yt-dlp>=2024.0.0
mutagen>=1.47.0
Pillow>=10.0.0
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
